+++ bot/apps.py
@@ -5,4 +5,4 @@ from django.utils.translation import gettext_lazy as _
 class BotConfig(AppConfig):
     default_auto_field = 'django.db.models.AutoField'
     name = 'bot'
-    verbose_name = _('Bot')
+    verbose_name = _('bot')
