--- bot/commands/roleplay/__init__.py
@@ -1,3 +1,0 @@
-from .places import WorldsCommand
-
-__all__ = ['WorldsCommand']
