--- bot/routing.py
@@ -1,7 +1,0 @@
-from django.urls import re_path
-
-from . import consumers
-
-websocket_urlpatterns = [
-    re_path(r'^ws/bot/register/$', consumers.BotConsumer.as_asgi(), name='register'),
-]
