+++ bot/urls.py
@@ -1,13 +1,6 @@
 from django.urls import include, path
-from rest_framework import routers
 
-from . import views, viewsets
-
-router = routers.DefaultRouter()
-router.register(r'discord_user', viewsets.DiscordUserViewSet)
-router.register(r'discord_server', viewsets.DiscordServerViewSet)
-router.register(r'discord_text_channel', viewsets.DiscordTextChannelViewSet)
-router.register(r'discord_voice_channel', viewsets.DiscordVoiceChannelViewSet)
+from . import views
 
 app_name = 'bot'
 
@@ -17,6 +10,5 @@ BOT_UTILS = [
 ]
 
 urlpatterns = [
-    path('api/', include(router.urls)),
     path('utils/', include((BOT_UTILS, app_name), namespace='utils')),
 ]
