+++ registration/forms.py
@@ -8,12 +8,11 @@ from django import forms
 from django.conf import settings
 from django.contrib.auth import forms as auth_forms
 from django.contrib.auth import get_user_model
-from django.contrib.auth.tokens import PasswordResetTokenGenerator
 from django.shortcuts import reverse
 from django.template.loader import render_to_string
 from django.utils.translation import gettext_lazy as _
 
-from bot.models import DiscordUser
+from common.utils.auth import generate_token
 
 LOGGER = logging.getLogger(__name__)
 
@@ -169,25 +168,6 @@ class SignUpForm(auth_forms.UserCreationForm):
         for field in required_fields:
             self.fields[field].required = True
 
-    def _generate_token(self, user) -> str:
-        """
-        Generates a token for the user to confirm it's email.
-
-        Parameters
-        ----------
-        user: User instance
-            The user associated to this token.
-
-        Returns
-        -------
-        token: :class:`str`
-            The token generated.
-        """
-
-        token = PasswordResetTokenGenerator()
-        token = token.make_token(user)
-        return token
-
     def _send_email_confirmation(self, user):
         """
         Sends a confirmation email to the user.
@@ -196,7 +176,7 @@ class SignUpForm(auth_forms.UserCreationForm):
         msg_html = render_to_string('email_templates/confirm_email.html', {
             # We declare localhost as default for tests purposes
             'domain': self.request.META.get('HTTP_HOST', 'http://localhost'),
-            'token': self._generate_token(user),
+            'token': generate_token(user),
             'object': user
         })
 
@@ -204,30 +184,6 @@ class SignUpForm(auth_forms.UserCreationForm):
             user.email_user(_('Welcome to Oil & Rope!'), '', html_message=msg_html)
         except SMTPAuthenticationError:  # pragma: no cover
             LOGGER.exception('Unable to logging email server with given credentials.')
-
-    def clean_discord_id(self):
-        """
-        Checks if Discord User is created.
-        """
-
-        data = self.cleaned_data.get('discord_id')
-
-        if data:
-            if not DiscordUser.objects.filter(pk=data).exists():
-                msg = '{}. {}'.format(_('User not found'), _('Have you requested invitation?'))
-                self.add_error('discord_id', msg)
-        return data
-
-    def get_discord_user(self):
-        """
-        Looks for `discord_id` field and returns :class:`DiscordUser` instance or `None`.
-        """
-
-        discord_user = None
-        discord_id = self.cleaned_data.get('discord_id')
-        if discord_id:
-            discord_user = DiscordUser.objects.get(pk=discord_id)
-        return discord_user
 
     def save(self, commit=True):
         """
@@ -242,14 +198,8 @@ class SignUpForm(auth_forms.UserCreationForm):
         instance = super().save(commit=False)
         # Set active to False until user activates email
         instance.is_active = False
-        # Checks for DiscordUser
-        discord_user = self.get_discord_user()
         if commit:
             instance.save()
-            # Adds foreign key if exists
-            if discord_user:
-                discord_user.user = instance
-                discord_user.save()
             # User shouldn't wait for the email to be sent
             with ThreadPoolExecutor(max_workers=2) as executor:
                 executor.submit(self._send_email_confirmation, instance)
