--- tests/bot/test_consumers.py
@@ -1,70 +1,0 @@
-"""
-:class:`unittest.TestCase` are not compatible with WebSockets Consumers (DjangoChannels) so tests must
-written as top-level functions.
-"""
-
-
-import faker
-import pytest
-from channels.db import database_sync_to_async as sync_to_async
-from channels.testing import WebsocketCommunicator
-from django.shortcuts import reverse
-from model_bakery import baker
-
-from bot import models
-from bot.consumers import BotConsumer
-
-fake = faker.Faker()
-
-
-@pytest.fixture(scope='function', autouse=False)
-def url():
-    return reverse('registration:register')
-
-
-@pytest.mark.asyncio
-async def test_websocket_connect(url):
-    communicator = WebsocketCommunicator(BotConsumer.as_asgi(), url)
-    connected, subprotocol = await communicator.connect()
-    assert connected, 'WebSocket is not connected.'
-
-    # Disconnect WebSocket to avoid pending tasks
-    await communicator.disconnect()
-
-
-@pytest.mark.django_db
-@pytest.mark.asyncio
-async def test_discord_user_does_not_exist(url):
-    communicator = WebsocketCommunicator(BotConsumer.as_asgi(), url)
-    data = {
-        'type': 'check_user',
-        'discord_id': fake.random_int()
-    }
-    await communicator.send_json_to(data)
-    response = await communicator.receive_json_from()
-    assert 'exists' in response, 'Key \'exists\' does not exist.'
-    assert not response['exists'], 'Discord ID check as existent.'
-
-    # Disconnect WebSocket to avoid pending tasks
-    await communicator.disconnect()
-
-
-@pytest.mark.django_db
-@pytest.mark.asyncio
-async def test_discord_user_exists(url):
-    discord_user = await sync_to_async(baker.make)(models.DiscordUser)
-    communicator = WebsocketCommunicator(BotConsumer.as_asgi(), url)
-
-    data = {
-        'type': 'check_user',
-        'discord_id': discord_user.id
-    }
-    await communicator.send_json_to(data)
-    response = await communicator.receive_json_from()
-
-    assert 'exists' in response, 'Key \'exists\' does not exist.'
-    assert response['exists'], 'Discord ID check as non existent.'
-
-    # Disconnect WebSocket to avoid pending tasks
-    await communicator.disconnect()
-    await sync_to_async(discord_user.delete)()
