+++ tests/registration/test_models.py
@@ -1,6 +1,6 @@
 from django.apps import apps
 from django.contrib.auth import get_user_model
-from django.test import TestCase, override_settings
+from django.test import TestCase
 from django.utils import timezone
 from faker import Faker
 from freezegun import freeze_time
@@ -8,7 +8,6 @@ from model_bakery import baker
 
 from common.constants import models as constants
 from roleplay.enums import SiteTypes
-from tests.bot.helpers.constants import LITECORD_API_URL, LITECORD_TOKEN, USER_WITH_SAME_SERVER
 
 Place = apps.get_model(constants.PLACE_MODEL)
 Race = apps.get_model(constants.RACE_MODEL)
@@ -24,22 +23,6 @@ class TestUser(TestCase):
         cls.model = get_user_model()
 
         cls.instance = baker.make(_model=cls.model)
-        cls.discord_user = baker.make(
-            _model=constants.DISCORD_USER_MODEL,
-            id=USER_WITH_SAME_SERVER,
-            user=cls.instance,
-        )
-
-    def test_get_user_from_discord_api_ko(self):
-        user = baker.make(_model=self.model)
-        discord_user = user.get_user_from_discord_api()
-
-        self.assertIsNone(discord_user)
-
-    @override_settings(DISCORD_API_URL=LITECORD_API_URL, BOT_TOKEN=LITECORD_TOKEN)
-    def test_get_user_from_discord_api_ok(self):
-        discord_user = self.instance.get_user_from_discord_api()
-        self.assertIsNotNone(discord_user)
 
     def test_owned_races_ok(self):
         iterations = fake.pyint(min_value=1, max_value=10)
